# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EnergieGo API",
        "description": "Documentation OpenAPI pour l'API EnergieGo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/administrateurs": {
            "get": {
                "summary": "Lister les administrateurs",
                "operationId": "ab08cb936a9c745af6af02e581336d38",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Administrateur"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer un administrateur",
                "operationId": "5f8954fef358d91cffc5bf0c90d8d873",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Administrateur"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Administrateur"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/administrateurs/{id}": {
            "get": {
                "summary": "Afficher un administrateur",
                "operationId": "e0526ee2b972c3e30bbe0cb8da3fc0d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Administrateur"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un administrateur",
                "operationId": "fba8d818329d80eabba3e71b829373cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Administrateur"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Administrateur"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un administrateur",
                "operationId": "3fee819365071db452cdf273ca697043",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
        "/api/avis": {
            "get": {
                "summary": "Lister les avis",
                "operationId": "074b394984977e6dba6974626b17ff22",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Avis"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer un avis",
                "operationId": "12b76ddbc444262178f834957435877d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Avis"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avis"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/avis/{id}": {
            "get": {
                "summary": "Afficher un avis",
                "operationId": "4167dc39a58a676e2b99040ac98ae715",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avis"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un avis",
                "operationId": "06de6748d7152d730e65e38ef8816689",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Avis"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avis"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un avis",
                "operationId": "c158d9539417f8b9af2ad21eec4d7f98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "summary": "Lister les catégories",
                "operationId": "0a3810ef08509327a6db442d3e4c4b73",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Categorie"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer une catégorie",
                "operationId": "f44066b315f1a6670d352e128ba39465",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorie"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "summary": "Afficher une catégorie",
                "operationId": "829dd0fbc1687efece34524110808262",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour une catégorie",
                "operationId": "68cf6ef85d9b79cea99f6063deca74e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorie"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "summary": "Supprimer une catégorie",
                "operationId": "c637fcbceb0196c4d850f96245587511",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Categorie": {
                "required": [
                    "nom",
                    "description"
                ],
                "properties": {
                    "id_categorie": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Administrateur": {
                "required": [
                    "id_utilisateur"
                ],
                "properties": {
                    "id_utilisateur": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Avis": {
                "required": [
                    "id_produit",
                    "id_utilisateur",
                    "note",
                    "commentaire",
                    "dateAvis"
                ],
                "properties": {
                    "id_avis": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_produit": {
                        "type": "integer"
                    },
                    "id_utilisateur": {
                        "type": "integer"
                    },
                    "note": {
                        "type": "integer"
                    },
                    "commentaire": {
                        "type": "string"
                    },
                    "dateAvis": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}